Get rid of Nat Gateway
This reconfigures the VPC to only have a single public subnet rather
than a private and a public one, and moves the service into that public
subnet. This allows us to get rid of the NAT Gateway that otherwise accrues
cost for no real reason.

The service still needs to be in a public subnet in order to have internet
access (for the Confluent Cloud cluster). It's necessary to assign public
IP addresses to the instances in order to be able to pull from the Docker registry.